changes based on comments #5364
